Integrated the headerPressable into the PressableButton, and created the writetodb firebase helper function.
